Student specific feedback
You aren't checking for uniqueness on your responses - I know you're not allowing them to be non-unique based on what you're passing in, but for this kind of requirement it would be better to do some checking to validate the passed in responses
It's good to see bounds checks done in a setter
Remember foreach loops - they're much cleaner and therefore easier to maintain, when they can be used.
Make sure you name your methods and variables what they do/are (for example, the getLastResponse method didn't say what it really did, before). There's no need to get clever with your code to make sure you call the same method with different parameters to do two different things.