GraphicalChat.java feedback
Setting up your elements in methods is not a bad idea to keep your calls separated. It is possibly even better to set your listeners up separately as well, but those are decisions that need to be vetted per project, of course.
Note that your fonts should both be constants (and are, the way you are using them). Therefore, more appropriate uses would be to set up the name as HELVETICA_FONT (instead of hv) and your font as a static final (as you have to do with serialVersionUID).